requirements.txt
# Layer-1 Dependencies (Data Collection & Ingestion)
anvil-uplink
PyGithub
boto3
python-dotenv
requests

# Layer-2 Dependencies (Intelligence & Analysis)
scikit-learn>=1.0.0
pandas>=1.3.0
numpy>=1.21.0
nltk>=3.6.0
